data count use long

+++ src/Jane/Application/Services/Dto/IHasTotalCount.cs
@@ -8,6 +8,6 @@ namespace Jane.Application.Services.Dto
         /// <summary>
         /// Total count of Items.
         /// </summary>
-        int TotalCount { get; set; }
+        long TotalCount { get; set; }
     }
 }
